[EUREKA-561-3]. Add validateUserUnaffiliatedLocationUpdates method to Order Open/unopen
